Bugs fixed

+++ src/main/java/com/imaginnovate/demo/repository/BatchProcessRepository.java
@@ -4,4 +4,4 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.imaginnovate.demo.entity.BatchProcess;
 
-public interface BatchProcessRepository extends JpaRepository<BatchProcess, Integer> { }
+public interface BatchProcessRepository extends JpaRepository<BatchProcess, Long> { }
